Swap water level and contour interval to doubles (totally blanked on the meters issue)

+++ SandWorm/Analytics/Contours.cs
@@ -9,7 +9,7 @@ namespace SandWorm.Analytics
         {
         }
 
-        public override void GetGeometryForAnalysis(ref List<GeometryBase> outputGeometry, int contourInterval, Mesh mesh)
+        public override void GetGeometryForAnalysis(ref List<GeometryBase> outputGeometry, double contourInterval, Mesh mesh)
         {
             var bounds = mesh.GetBoundingBox(false);
             var originStart = new Point3d(0, 0, bounds.Min.Z);
